requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

xgboost==2.1.1

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

pydantic==2.8.2
python-dotenv==1.0.1

joblib==1.4.2
tqdm==4.66.4

matplotlib==3.9.1
statsmodels==0.14.2
